+++ airslate/entities/tags.py
@@ -5,15 +5,15 @@
 # For the full copyright and license information, please view
 # the LICENSE file that was distributed with this source code.
 
-"""Provides Document Fields related classes and functionality."""
+"""Provides Tags classes and functionality."""
 
 
 from .base import BaseEntity
 
 
-class Field(BaseEntity):
-    """Represents Document Fields entity."""
+class Tag(BaseEntity):
+    """Represents Tags entity."""
 
     @property
     def type(self):
-        return 'dictionary'
+        return 'flow_tags'